fix: preserve coordinate precision in CMS

+++ server/models.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import JSON, Boolean, DateTime, Float, ForeignKey, Index, Integer, String, Text, UniqueConstraint
+from sqlalchemy import JSON, Boolean, DateTime, Double, Float, ForeignKey, Index, Integer, String, Text, UniqueConstraint
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 
@@ -60,8 +60,8 @@ class City(Base):
     name: Mapped[str] = mapped_column(String(80))
     subtitle: Mapped[str] = mapped_column(String(160))
     hero_image: Mapped[str] = mapped_column(String(255))
-    latitude: Mapped[float] = mapped_column(Float)
-    longitude: Mapped[float] = mapped_column(Float)
+    latitude: Mapped[float] = mapped_column(Double)
+    longitude: Mapped[float] = mapped_column(Double)
 
 
 class Route(Base):
@@ -94,8 +94,8 @@ class Stop(Base):
     title: Mapped[str] = mapped_column(String(160))
     kicker: Mapped[str] = mapped_column(String(120))
     address: Mapped[str] = mapped_column(String(255))
-    latitude: Mapped[float] = mapped_column(Float)
-    longitude: Mapped[float] = mapped_column(Float)
+    latitude: Mapped[float] = mapped_column(Double)
+    longitude: Mapped[float] = mapped_column(Double)
     arrival_radius_m: Mapped[int] = mapped_column(Integer, default=80)
     story_title: Mapped[str] = mapped_column(String(200))
     story_body: Mapped[str] = mapped_column(Text)
@@ -359,8 +359,8 @@ class TriggerRegion(Base):
 
     id: Mapped[str] = mapped_column(String(36), primary_key=True)
     fragment_id: Mapped[str] = mapped_column(String(36), unique=True)
-    latitude: Mapped[float] = mapped_column(Float)
-    longitude: Mapped[float] = mapped_column(Float)
+    latitude: Mapped[float] = mapped_column(Double)
+    longitude: Mapped[float] = mapped_column(Double)
     entry_radius_m: Mapped[int] = mapped_column(Integer, default=60)
     exit_radius_m: Mapped[int] = mapped_column(Integer, default=90)
     max_accuracy_m: Mapped[int] = mapped_column(Integer, default=50)
